+++ inherited-rules/unused/import.js
@@ -1,0 +1,8 @@
+"use strict";
+module.exports = {
+  rules: {
+    "import/no-named-default": "error",
+    "import/first": "error",
+    "import/no-webpack-loader-syntax": "error"
+  }
+};